Add dependency to systemctl and chkconfig processors (#558)
Also, combine duplicate Processor classes.
